Rust generator: Fix type conversion when creating a Point
Before it used to always be a property access, but now it may be inlined
and result in a f64 or int expression

+++ internal/compiler/generator/rust.rs
@@ -1733,7 +1733,7 @@ fn compile_builtin_function_call(
                 quote!(
                     #window_tokens.show_popup(
                         &VRc::into_dyn(#popup_window_id::new(_self.self_weak.get().unwrap().clone()).into()),
-                        Point::new(#x, #y),
+                        Point::new(#x as f32, #y as f32),
                         #parent_component
                     );
                 )
